fix: update published date

+++ app/Repositories/GalleryRepository.php
@@ -25,13 +25,11 @@ class GalleryRepository implements GalleryInterface
     {
         $fileNameImage = uniqid() . '.' . $data['image']->extension();
         $data['image']->storeAs('public/gallery', $fileNameImage);
+        $data['image'] = $fileNameImage;
 
         DB::beginTransaction();
         try {
-            $this->gallery->create([
-                'image' => $fileNameImage,
-                'title' => $data['title'],
-            ]);
+            $this->gallery->create($data);
             DB::commit();
         } catch (\Throwable $th) {
             Storage::delete('public/gallery/' . $fileNameImage);
@@ -44,16 +42,16 @@ class GalleryRepository implements GalleryInterface
     {
         $gallery = $this->gallery->find($id);
 
-        Storage::delete('public/gallery/' . $gallery->image);
-        $fileNameImage = uniqid() . '.' . $data['image']->extension();
-        $data['image']->storeAs('public/gallery', $fileNameImage);
+        if (isset($data['image'])) {
+            Storage::delete('public/gallery/' . $gallery->image);
+            $fileNameImage = uniqid() . '.' . $data['image']->extension();
+            $data['image']->storeAs('public/gallery', $fileNameImage);
+            $data['image'] = $fileNameImage;
+        }
 
         DB::beginTransaction();
         try {
-            $gallery->update([
-                'image' => $fileNameImage,
-                'title' => $data['title'],
-            ]);
+            $gallery->update($data);
             DB::commit();
         } catch (\Throwable $th) {
             Storage::delete('public/gallery/' . $fileNameImage);
@@ -86,7 +84,5 @@ class GalleryRepository implements GalleryInterface
     public function countGallery()
     {
         return $this->gallery->get()->count();
-
     }
-
 }
